+++ Godeps/_workspace/src/github.com/jbenet/go-multihash/sum.go
@@ -7,7 +7,7 @@ import (
 	"errors"
 	"fmt"
 
-	sha3 "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.crypto/sha3"
+	sha3 "github.com/jbenet/go-ipfs/Godeps/_workspace/src/golang.org/x/crypto/sha3"
 )
 
 var ErrSumNotSupported = errors.New("Function not implemented. Complain to lib maintainer.")
