+++ Godeps/_workspace/src/golang.org/x/crypto/sha3/register.go
