+++ Godeps/_workspace/src/golang.org/x/crypto/sha3/testdata/keccakKats.json.deflate
